requirements.txt
ipwhois==1.2.0
more-itertools==10.2.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-geoip==1.2
python-retry==0.0.1
pywin32==306
requests==2.31.0